+++ R/connections.R
@@ -18,19 +18,25 @@
 #'
 #' @param ... Additional information to be added to the final data output
 #'
+#' @param weight An expression to be evaluated on the edge data to provide
+#' weigths for the shortest path calculations
+#'
+#' @inheritParams igraph::shortest_paths
+#'
 #' @return A function that takes a layout_ggraph object and returns the given
 #' connections
 #'
 #' @family extractors
 #'
 #' @export
-get_con <- function(from = integer(), to = integer(), paths = NULL, ...) {
+get_con <- function(from = integer(), to = integer(), paths = NULL, ..., weight = NULL, mode = 'all') {
     if (length(from) != length(to)) {
         stop('from and to must be of equal length')
     }
     function(layout) {
         if (length(from) == 0) return(NULL)
-        connections <- getConnections(layout, from, to)
+        connections <- getConnections(layout = layout, from = from, to = to,
+                                      weight = {{ weight }}, mode = mode)
         nodes <- as.data.frame(layout)[unlist(connections), ]
         nodes$con.id <- rep(seq_along(connections), lengths(connections))
         if (!is.null(paths)) {